app/requirements.txt
fastapi
ldap3
pyjwt
zeep
